add README and slight lateration to clear local storage

+++ comp6000/test/comp6000/contexts/storage_test.exs
@@ -31,15 +31,17 @@ defmodule Comp6000.Contexts.StorageTest do
   # An exceedingly nasty function to delete all files and directories within local-storage once tests are complete
   defp clear_local_storage() do
     Enum.map(File.ls!("#{@storage_path}"), fn study_dir ->
-      Enum.map(File.ls!("#{@storage_path}/#{study_dir}"), fn task_dir ->
-        Enum.map(File.ls!("#{@storage_path}/#{study_dir}/#{task_dir}"), fn file ->
-          File.rm("#{@storage_path}/#{study_dir}/#{task_dir}/#{file}")
+      if File.dir?(study_dir) do
+        Enum.map(File.ls!("#{@storage_path}/#{study_dir}"), fn task_dir ->
+          Enum.map(File.ls!("#{@storage_path}/#{study_dir}/#{task_dir}"), fn file ->
+            File.rm("#{@storage_path}/#{study_dir}/#{task_dir}/#{file}")
+          end)
+
+          File.rmdir!("#{@storage_path}/#{study_dir}/#{task_dir}")
         end)
 
-        File.rmdir!("#{@storage_path}/#{study_dir}/#{task_dir}")
-      end)
-
-      File.rmdir!("#{@storage_path}/#{study_dir}")
+        File.rmdir!("#{@storage_path}/#{study_dir}")
+      end
     end)
   end
